fix(broker/bam): big metrics do not fail anymore when inserted in database
* fix(broker): new script to compute database columns max sizes

* enh(broker): table_max_size.py restored

* fix(broker/bam): the good size of float is inserted into the database for bam

* fix(broker): perfdata class contains float numbers now.

* enh(tests): big metrics do not fail anymore when inserted in database

* fix(tests): a syntax error fixed

* fix(broker): changes after review

* fix(broker): service_status forgotten

* cleanup(tests): comments removed

* fix(broker): double to float changes the precision

* fix(broker): some doubles forgotten

REFS: MON-116620

+++ broker/storage/inc/com/centreon/broker/storage/stream.hh
@@ -52,22 +52,22 @@ class stream : public io::stream {
     bool locked;
     uint32_t metric_id;
     uint16_t type;
-    double value;
+    float value;
     std::string unit_name;
-    double warn;
-    double warn_low;
+    float warn;
+    float warn_low;
     bool warn_mode;
-    double crit;
-    double crit_low;
+    float crit;
+    float crit_low;
     bool crit_mode;
-    double min;
-    double max;
+    float min;
+    float max;
   };
   struct metric_value {
     time_t c_time;
     uint32_t metric_id;
     short status;
-    double value;
+    float value;
   };
 
   std::string _status;
